# Add some primitives; speed up tests