--- a/inferno-core/test/Eval/Spec.hs
+++ b/inferno-core/test/Eval/Spec.hs
@@ -7,7 +7,7 @@ import Data.Bifunctor (bimap)
 import Data.Int (Int64)
 import qualified Data.Map as Map
 import Data.Text (unpack)
-import Inferno.Core (Interpreter (evalInEnv, parseAndInfer, parseAndInferTypeReps), mkInferno)
+import Inferno.Core (Interpreter (evalInEnv, parseAndInfer, parseAndInferTypeReps, evalInEmptyEnv), mkInferno)
 import Inferno.Eval.Error (EvalError (..))
 import Inferno.Module.Builtin (enumBoolHash)
 import qualified Inferno.Module.Prelude as Prelude
@@ -15,12 +15,12 @@ import Inferno.Types.Syntax (BaseType (..), Expr (..), ExtIdent (..), Ident (..)
 import Inferno.Types.Value (Value (..))
 import Inferno.Types.VersionControl (pinnedToMaybe)
 import Inferno.Utils.Prettyprinter (renderPretty)
-import Test.Hspec (Spec, describe, expectationFailure, it, shouldBe)
+import Test.Hspec (Spec, describe, expectationFailure, it, shouldBe, runIO)
 
 type TestCustomValue = ()
 
-runtimeTypeRepsTests :: Spec
-runtimeTypeRepsTests = describe "runtime type reps" $ do
+runtimeTypeRepsTests :: Interpreter TestCustomValue -> Spec
+runtimeTypeRepsTests inferno = describe "runtime type reps" $ do
   let expr_3 =
         case parseAndInfer inferno "3" of
           Left err ->
@@ -38,7 +38,6 @@ runtimeTypeRepsTests = describe "runtime type reps" $ do
     let expr_3_double = App expr (TypeRep () (TBase TDouble))
     shouldEvaluateTo expr_3_double (VDouble 3)
   where
-    inferno = mkInferno Prelude.builtinModules :: Interpreter TestCustomValue
     shouldEvaluateTo expr (v :: Value TestCustomValue IO) = do
       evalInEnv inferno Map.empty Map.empty expr >>= \case
         Left err -> expectationFailure $ "Failed eval with: " <> show err
@@ -47,6 +46,36 @@ runtimeTypeRepsTests = describe "runtime type reps" $ do
 evalTests :: Spec
 evalTests = describe "evaluate" $
   do
+    inferno <- runIO $ (mkInferno Prelude.builtinModules :: IO (Interpreter TestCustomValue))
+    let shouldEvaluateInEnvTo localEnv implEnv str (v :: Value TestCustomValue IO) =
+          it ("\"" <> unpack str <> "\" should evaluate to " <> (unpack $ renderPretty v)) $ do
+            case parseAndInferTypeReps inferno str of
+              Left err -> expectationFailure $ show err
+              Right ast ->
+                evalInEnv inferno localEnv implEnv ast >>= \case
+                  Left err -> expectationFailure $ "Failed eval with: " <> show err
+                  Right v' -> (renderPretty v') `shouldBe` (renderPretty v)
+    let shouldEvaluateTo str (v :: Value TestCustomValue IO) =
+          it ("\"" <> unpack str <> "\" should evaluate to " <> (unpack $ renderPretty v)) $ do
+            case parseAndInferTypeReps inferno str of
+              Left err -> expectationFailure $ show err
+              Right ast ->
+                evalInEmptyEnv inferno Map.empty ast >>= \case
+                  Left err -> expectationFailure $ "Failed eval with: " <> show err
+                  Right v' -> (renderPretty v') `shouldBe` (renderPretty v)
+    let shouldThrowRuntimeError str merr =
+          it ("\"" <> unpack str <> "\" should throw a runtime error") $ do
+            case parseAndInferTypeReps inferno str of
+              Left err -> expectationFailure $ show err
+              Right ast ->
+                -- TODO fix implicit Env to empty too?
+                evalInEmptyEnv inferno Map.empty ast >>= \case
+                  Left err' -> case merr of
+                    Nothing -> pure ()
+                    Just err -> err' `shouldBe` err
+                  Right _ -> expectationFailure $ "Should not evaluate."
+
+
     shouldEvaluateTo "3" $ VDouble 3
     shouldEvaluateTo "-3" $ VDouble (-3)
     shouldEvaluateTo "-(-3)" $ VDouble 3
@@ -376,27 +405,7 @@ evalTests = describe "evaluate" $
     shouldThrowRuntimeError "assert #false in ()" $ Just AssertionFailed
     shouldEvaluateTo "assert #true in ()" $ VTuple []
 
-    runtimeTypeRepsTests
+    runtimeTypeRepsTests inferno
   where
     vTrue = VEnum enumBoolHash (Ident "true")
     vFalse = VEnum enumBoolHash (Ident "false")
-    inferno = mkInferno Prelude.builtinModules :: Interpreter TestCustomValue
-    shouldEvaluateInEnvTo localEnv implEnv str (v :: Value TestCustomValue IO) =
-      it ("\"" <> unpack str <> "\" should evaluate to " <> (unpack $ renderPretty v)) $ do
-        case parseAndInferTypeReps inferno str of
-          Left err -> expectationFailure $ show err
-          Right ast ->
-            evalInEnv inferno localEnv implEnv ast >>= \case
-              Left err -> expectationFailure $ "Failed eval with: " <> show err
-              Right v' -> (renderPretty v') `shouldBe` (renderPretty v)
-    shouldEvaluateTo = shouldEvaluateInEnvTo Map.empty Map.empty
-    shouldThrowRuntimeError str merr =
-      it ("\"" <> unpack str <> "\" should throw a runtime error") $ do
-        case parseAndInferTypeReps inferno str of
-          Left err -> expectationFailure $ show err
-          Right ast ->
-            evalInEnv inferno Map.empty Map.empty ast >>= \case
-              Left err' -> case merr of
-                Nothing -> pure ()
-                Just err -> err' `shouldBe` err
-              Right _ -> expectationFailure $ "Should not evaluate."