This PR adds some (long awaited) primitives to Inferno. :)

```ocaml
// Function composition with `<<`, `.` creates parsing issues
assert Array.map ((Text.append "a") << (Text.append "b")) ["0", "1"] == ["ab0", "ab1"] in
// Pipe operator `|>`
assert "0" |> Text.append "a" |> Text.append "b" == "ba0" in
// Some tuple stuff:
assert fst (1, 0) == snd (0, 1) in
assert zip [1, 2, 3] [4, 5] == [(1, 4), (2, 5)] in
assert zip [1, 2] ["a", "b"] == [(1,"a"),(2,"b")] in
assert zip [1] ["a", "b"] == [(1,"a")] in
assert zip [1, 2] ["a"] == [(1,"a")] in
assert zip [] [1, 2] == [] in
assert zip [1, 2] [] == [] in
0
```

While adding these, I noticed that the Eval tests were very slow because they were re-evaluating the prelude for each test. This PR also adds a new function `evalInEmptyEnv` that shares this computation and speeds up the tests by 3x. Before:

```
HSPEC_PRINT_SLOW_ITEMS=10 cabal test inferno-tests  --test-option=--match --test-option="evaluate" --test-show-details=direct
...
Finished in 16.6665 seconds
262 examples, 0 failures

Slow spec items:
  test/Eval/Spec.hs:375:7: /evaluate/"match [1.2, 3, 3] with { | [x, y, z] -> 2*x+3*y+z | _ -> 3 }" should evaluate to 14.4/ (80ms)
  test/Eval/Spec.hs:375:7: /evaluate/"(Array.reduce (fun x y -> x + max 0 y) 0 ((-3) .. 3)) == (Array.reduceRight (fun x y -> y + max 0 x) 0 ((-3) .. 3))" should evaluate to #true/ (80ms)
  test/Eval/Spec.hs:375:7: /evaluate/"open Time in let ?now = toTime (seconds 4000) in intervalEvery (seconds 4) ?now (?now + seconds 10)" should evaluate to [4000s,4004s,4008s]/ (79ms)
  test/Eval/Spec.hs:375:7: /evaluate/"open Time in let ?now = (toTime (weeks 66666)) in weeksBefore ?now 44 == ?now - (weeks 44)" should evaluate to #true/ (77ms)
  test/Eval/Spec.hs:375:7: /evaluate/"[x | x <- 1 .. 10]" should evaluate to [1,2,3,4,5,6,7,8,9,10]/ (75ms)
  test/Eval/Spec.hs:375:7: /evaluate/"open Time in let ?now = (toTime (hours 66666)) in hoursBefore ?now 44 == ?now - (hours 44)" should evaluate to #true/ (75ms)
  test/Eval/Spec.hs:375:7: /evaluate/"open Time in let ?now = (toTime (minutes 66666)) in minutesBefore ?now 44 == ?now - (minutes 44)" should evaluate to #true/ (75ms)
  test/Eval/Spec.hs:375:7: /evaluate/"Array.findFirstAndLastSome [None, Some 3.0, None, Some 4.0]" should evaluate to Some (3.0, 4.0)/ (75ms)
  test/Eval/Spec.hs:375:7: /evaluate/"open Time in let ?now = (toTime (seconds 66666)) in secondsBefore ?now 44 == ?now - (seconds 44)" should evaluate to #true/ (74ms)
  test/Eval/Spec.hs:375:7: /evaluate/"open Time in day (toTime (days 3 + hours 22)) == toTime (days 3)" should evaluate to #true/ (74ms)
Test suite inferno-tests: PASS
```
After:
```
HSPEC_PRINT_SLOW_ITEMS=10 cabal test inferno-tests  --test-option=--match --test-option="evaluate" --test-show-details=direct
...
Finished in 5.7441 seconds
271 examples, 0 failures

Slow spec items:
  test/Eval/Spec.hs:51:11: /evaluate/"?x + 2" should evaluate to 7.0/ (1058ms)
  test/Eval/Spec.hs:31:3: /evaluate/runtime type reps/test int type rep/ (1037ms)
  test/Eval/Spec.hs:51:11: /evaluate/"let f = fun x -> ?x + 2 in f 0" should evaluate to 7.0/ (1037ms)
  test/Eval/Spec.hs:36:3: /evaluate/runtime type reps/test double type rep/ (1017ms)
  test/Eval/Spec.hs:59:11: /evaluate/"Option.reduce (fun d -> d + 2) 0 (Some 4.0)" should evaluate to 6.0/ (61ms)
  test/Eval/Spec.hs:59:11: /evaluate/"Array.getOpt [0, 1, 2] (-9)" should evaluate to None/ (59ms)
  test/Eval/Spec.hs:59:11: /evaluate/"Array.average [0.0, 1.0]" should evaluate to 0.5/ (58ms)
  test/Eval/Spec.hs:59:11: /evaluate/"zip [1, 2] ["a", "b"] == [(1,"a"),(2,"b")]" should evaluate to #true/ (26ms)
  test/Eval/Spec.hs:59:11: /evaluate/"zip [1, 2, 3] [4, 5] == [(1, 4), (2, 5)]" should evaluate to #true/ (24ms)
  test/Eval/Spec.hs:59:11: /evaluate/"zip [1, 2] ["a"] == [(1,"a")]" should evaluate to #true/ (21ms)
Test suite inferno-tests: PASS
```

Note that the test `/evaluate/"?x + 2"` is taking much longer than before, this is because it uses a non-empty implicit env so cannot benefit from the `evaluateInEmptyEnv` speedup. We need to profile this execution and see why adding these primitives makes the prelude evaluation so much slower.