@Daniel-Diaz 2023-09-13
I think a solution that should work for both empty and non-empty environments is being able to pass such environment to `mkInferno`, instead of being generated by it. This would make the run-with-empty-environment function redundant and all evaluations will get the optimization.