@siddharth-krishna 2023-09-19
@Daniel-Diaz I split out the refactor into a separate PR #82 so now this PR is just the new primitives. Let me know if it looks good and we can merge this one in too?